LaserOperation made into a node type
